+++ dump_weights.py
@@ -10,4 +10,4 @@ L=[0.81822322 ,0.58561471 ,0.44783572 ,0.57275441 ,0.09763068 ,0.14935456 ,0.626
 #L=[0.8212428  ,0.58401753 ,0.8640102  ,0.99555712 ,0.01810704 ,0.14626944  ,0.61706742 ,0.54840212 ,0.30053032 ,0.05301519]
 L=np.array(L)
 with open('weights.pkl','wb') as f:
-                pickle.dump(L,f)
+    pickle.dump(L,f)
