--- full_balanced_run_output
@@ -1,19 +1,0 @@
-==========================================
-SLURM_CLUSTER_NAME = param-shivay
-SLURM_JOB_ACCOUNT = comp_sci_engg
-SLURM_JOB_ID = 729836
-SLURM_JOB_NAME = balanced_accuracy_script.sh
-SLURM_JOB_NODELIST = cn154
-SLURM_JOB_USER = parastiwari.rs.cse19.iitbhu
-SLURM_JOB_UID = 5583
-SLURM_JOB_PARTITION = standard
-SLURM_TASK_PID = 349704
-SLURM_SUBMIT_DIR = /home/parastiwari.rs.cse19.iitbhu/GADLRuleBased
-SLURM_CPUS_ON_NODE = 40
-SLURM_NTASKS = 
-SLURM_TASK_PID = 349704
-==========================================
-Warning: METEOR is not configured
-CUDA is not available, fall back to CPU.
-Making log output file: exp/train_movie_dual_view_inc_seed_250.ml.copy.bi-directional.20220327-083827/output.log
-exp/train_movie_dual_view_inc_seed_250.ml.copy.bi-directional.20220327-083827
