+++ model/multi_view_multi_task_basic_seq2seq.py
@@ -228,110 +228,7 @@ class MultiViewMultiTaskBasicClassifySeq2Seq(nn.Module):
             enc_logit = enc_classifier_output
             enc_classify_attn_dist = None
             sent_enc_classify_attn_dist = None
-        # construct rating memory bank
-        # rating_memory_bank = None
-        # if self.rating_memory_pred:
-        #     # [5, rating_v_size, emb_size]
-        #     rating_memory_bank = self.encoder.embedding(self.rating_tokens_tensor)
-        #     if self.rating_memory_type == 'pred':
-        #         # [batch_size]
-        #         rating_select_idx = torch.argmax(enc_logit, dim=1, keepdim=False)
-        #     else:
-        #         assert self.rating_memory_type == 'gold'
-        #         # [batch_size]
-        #         rating_select_idx = rating
-        #     # [batch_size, rating_v_size, emb_size]
-        #     rating_memory_bank = rating_memory_bank.index_select(dim=0, index=rating_select_idx)
-        #     # apply rating bridge when configured
-        #     if self.rating_bridge is not None:
-        #         rating_memory_bank = self.rating_bridge(rating_memory_bank)
-        # Decoding
-        # h_t_init = self.init_decoder_state(word_encoder_final_state)  # [dec_layers, batch_size, decoder_size]
-        # max_target_length = trg.size(1)
-        #context = self.init_context(memory_bank)  # [batch, memory_bank_size]
-        # decoder_dist_all = []
-        # attention_dist_all = []
-        # sentiment_context_all = []
-        # if self.coverage_attn:
-        #     coverage = torch.zeros_like(src, dtype=torch.float).requires_grad_()  # [batch, max_src_seq]
-        #     #coverage_all = coverage.new_zeros((max_target_length, batch_size, max_src_len), dtype=torch.float)  # [max_trg_len, batch_size, max_src_len]
-        #     coverage_all = []
-        # else:
-        #     coverage = None
-        #     coverage_all = None
-        # if self.review_attn or self.dec_classify_input_type == 'dec_state':
-        #     decoder_memory_bank = h_t_init[-1, :, :].unsqueeze(1)  # [batch, 1, decoder_size]
-        #     assert decoder_memory_bank.size() == torch.Size([batch_size, 1, self.decoder_size])
-        # else:
-        #     decoder_memory_bank = None
-        # if self.orthogonal_loss:  # create a list of batch_size empty list
-        #     delimiter_decoder_states_2dlist = [[] for i in range(batch_size)]
-        # init y_t to be BOS token
-        # y_t_init = trg.new_ones(batch_size) * self.bos_idx  # [batch_size]
-        # for t in range(max_target_length):
-        #     # determine the hidden state that will be feed into the next step
-        #     # according to the time step or the target input
-        #     if t == 0:
-        #         h_t = h_t_init
-        #         y_t = y_t_init
-        #     else:
-        #         h_t = h_t_next
-        #         y_t = y_t_next
-        #     if (self.review_attn or self.dec_classify_input_type == 'dec_state') and t > 0:
-        #         decoder_memory_bank = torch.cat([decoder_memory_bank, h_t[-1, :, :].unsqueeze(1)], dim=1)  # [batch, t+1, decoder_size]
-        #     decoder_dist, h_t_next, _, sentiment_context, attn_dist, p_gen, coverage = \
-        #         self.decoder(y_t, h_t, memory_banks, src_masks, max_num_oov, src_oov, coverage, decoder_memory_bank, rating_memory_bank)
-        #     decoder_dist_all.append(decoder_dist.unsqueeze(1))  # [batch, 1, vocab_size]
-        #     attention_dist_all.append(attn_dist.unsqueeze(1))  # [batch, 1, src_seq_len]
-        #     if self.coverage_attn:
-        #         coverage_all.append(coverage.unsqueeze(1))  # [batch, 1, src_seq_len]
-        #     if sentiment_context is not None:
-        #         sentiment_context_all.append(sentiment_context)  # [batch_size, memory_bank_size]
-        #     y_t_next = trg[:, t]  # [batch]
-        #     # if this hidden state corresponds to the delimiter, stack it
-        #     if self.orthogonal_loss:
-        #         for i in range(batch_size):
-        #             if y_t_next[i].item() == self.sep_idx:
-        #                 delimiter_decoder_states_2dlist[i].append(h_t_next[-1, i, :])  # [decoder_size]
-        # # collect the hidden state of the final decoding step
-        # if self.dec_classify_input_type == 'dec_state':
-        #     decoder_memory_bank = torch.cat([decoder_memory_bank, h_t_next[-1, :, :].unsqueeze(1)], dim=1)  # [batch, max_target_length+1, decoder_size]
-        # decoder_dist_all = torch.cat(decoder_dist_all, dim=1)  # [batch_size, trg_len, vocab_size]
-        # attention_dist_all = torch.cat(attention_dist_all, dim=1)  # [batch_size, trg_len, src_len]
-        # # new
-        # if self.dec_classify_input_type == 'attn_vec':
-        #     assert len(sentiment_context_all) != 0
-        #     dec_classifier_input = torch.stack(sentiment_context_all, dim=1)  # [batch_size, trg_len, memory_bank_size]
-        # else:
-        #     assert self.dec_classify_input_type == 'dec_state'
-        #     dec_classifier_input = decoder_memory_bank[:, 1:, :]  # [batch_size, trg_len, memory_bank_size]
-        #     if self.detach_classify_dec_states:
-        #         dec_classifier_input = dec_classifier_input.detach()
-        # assert dec_classifier_input.size() == torch.Size((batch_size, max_target_length, self.num_directions * self.encoder_size))
-        # if self.coverage_attn:
-        #     coverage_all = torch.cat(coverage_all, dim=1)  # [batch_size, trg_len, src_len]
-        #     assert coverage_all.size() == torch.Size((batch_size, max_target_length, max_src_len))
-        # if self.copy_attn:
-        #     assert decoder_dist_all.size() == torch.Size((batch_size, max_target_length, self.vocab_size + max_num_oov))
-        # else:
-        #     assert decoder_dist_all.size() == torch.Size((batch_size, max_target_length, self.vocab_size))
-        # assert attention_dist_all.size() == torch.Size((batch_size, max_target_length, max_src_len))
-        # Pad delimiter_decoder_states_2dlist with zero vectors
-        # if self.orthogonal_loss:
-        #     assert len(delimiter_decoder_states_2dlist) == batch_size
-        #     delimiter_decoder_states_lens = [len(delimiter_decoder_states_2dlist[i]) for i in range(batch_size)]
-        #     # [batch_size, decoder_size, max_num_delimiters]
-        #     delimiter_decoder_states = self.tensor_2dlist_to_tensor(delimiter_decoder_states_2dlist, batch_size, self.decoder_size, delimiter_decoder_states_lens)
-        # else:
-        #     delimiter_decoder_states_lens = None
-        #     delimiter_decoder_states = None
-        # dec_classifier_output = self.dec_classifier(dec_classifier_input, trg_mask)
-        # if isinstance(dec_classifier_output, tuple):
-        #     dec_logit = dec_classifier_output[0]
-        #     dec_classify_attn_dist = dec_classifier_output[1][0]
-        # else:
-        #     dec_logit = dec_classifier_output
-        #     dec_classify_attn_dist = None
+
         logit = (enc_logit)
         classifier_attention_dist = (enc_classify_attn_dist, sent_enc_classify_attn_dist)
         return  word_encoder_final_state, logit, classifier_attention_dist
