--- try2vae_output
@@ -1,18 +1,0 @@
-==========================================
-SLURM_CLUSTER_NAME = param-shivay
-SLURM_JOB_ACCOUNT = material_sci_tech
-SLURM_JOB_ID = 724599
-SLURM_JOB_NAME = script.sh
-SLURM_JOB_NODELIST = gpu003
-SLURM_JOB_USER = parastiwari.rs.cse19.iitbhu
-SLURM_JOB_UID = 5583
-SLURM_JOB_PARTITION = standard
-SLURM_TASK_PID = 162357
-SLURM_SUBMIT_DIR = /home/parastiwari.rs.cse19.iitbhu/GADLRuleBased
-SLURM_CPUS_ON_NODE = 40
-SLURM_NTASKS = 
-SLURM_TASK_PID = 162357
-==========================================
-Warning: METEOR is not configured
-Making log output file: exp/train_movie_dual_view_inc_seed_250.ml.copy.bi-directional.20220315-020304/output.log
-exp/train_movie_dual_view_inc_seed_250.ml.copy.bi-directional.20220315-020304
